+++ digitaltwins/gen3/resources/version_1_2_3/element_mapping.xlsx
