+++ digitaltwins/gen3/resources/version_2_0_0/element_mapping.xlsx
